changes struct for dto commands
